added cart

+++ backend/model/user.go
@@ -1,7 +1,7 @@
 package model
 
 type User struct {
-	Id       int8   `json:"id"`
+	Id       int    `json:"id"`
 	Name     string `json:"name"`
 	Email    string `json:"email"`
 	Password string `json:"password"`
